+++ src/Position.hpp
@@ -5,9 +5,9 @@
 class Position {
 public:
 
-    int x;
-    int y;
-    Position(int x, int y) {
+    unsigned long int x;
+    unsigned long int y;
+    Position(unsigned long int x, unsigned long int y) {
 
         this->x = x;
         this->y = y;
